requirements.txt
piexif==1.1.3
Pillow==10.1.0
